Refactor: user -> member 로 이름 변경
