+++ src/main/java/gift/exception/user/MemberAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package gift.exception.user;
+
+public class MemberAlreadyExistsException extends RuntimeException {
+    public MemberAlreadyExistsException(String message) {
+        super(message);
+    }
+}
