+++ src/main/java/gift/exception/user/MemberNotFoundException.java
@@ -1,0 +1,7 @@
+package gift.exception.user;
+
+public class MemberNotFoundException extends RuntimeException {
+    public MemberNotFoundException(String message) {
+        super(message);
+    }
+}
