--- src/main/java/gift/exception/user/UserAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package gift.exception.user;
-
-public class UserAlreadyExistsException extends RuntimeException {
-    public UserAlreadyExistsException(String message) {
-        super(message);
-    }
-}
