--- src/main/java/gift/exception/user/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package gift.exception.user;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
